feat: add encrption and decrption utils in crypto package
