+++ server/internal/crypto/crypto.go
@@ -6,6 +6,8 @@ import (
 	"crypto/cipher"
 	"crypto/rand"
 	"crypto/sha256"
+	"encoding/base64"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -19,7 +21,7 @@ import (
 
 var (
 	kmsClient *kms.Client
-	keyId     string
+	keyID     string
 	localKey  []byte
 	useKMS    bool
 	once      sync.Once
@@ -38,7 +40,7 @@ func InitEncryption() error {
 				return
 			}
 			kmsClient = kms.NewFromConfig(cfg)
-			keyId = key
+			keyID = key
 			useKMS = true
 		} else {
 			// Local AES-GCM Mode with SHA-256 derived key
@@ -57,7 +59,7 @@ func Encrypt(plaintext string) ([]byte, error) {
 	}
 	if useKMS {
 		out, err := kmsClient.Encrypt(context.Background(), &kms.EncryptInput{
-			KeyId:     &keyId,
+			KeyId:     &keyID,
 			Plaintext: []byte(plaintext),
 		})
 		if err != nil {
@@ -122,4 +124,55 @@ func Decrypt(cipherData []byte) (string, error) {
 	}
 
 	return string(plaintext), nil
+}
+
+type cryptoObj struct {
+	EncryptedData string `json:"encrypted_data"`
+}
+
+// EncryptJSONString encrypts the entire JSON string as a single value
+func EncryptJSONString(rawConfig string) (string, error) {
+
+	// Encrypt the entire config string
+	encryptedBytes, err := Encrypt(rawConfig)
+	if err != nil {
+		return "", fmt.Errorf("encryption failed: %v", err)
+	}
+	cryptoObj := cryptoObj{}
+
+	// Create a structured object with the encrypted data
+	cryptoObj.EncryptedData = base64.StdEncoding.EncodeToString(encryptedBytes)
+
+	// Marshal to JSON
+	encryptedJSON, err := json.Marshal(cryptoObj)
+	if err != nil {
+		return "", fmt.Errorf("failed to marshal encrypted data: %v", err)
+	}
+
+	return string(encryptedJSON), nil
+}
+
+// DecryptJSONObject decrypts a JSON object in the format {"encrypted_data": "base64-encoded-encrypted-json"}
+// and returns the original JSON string
+func DecryptJSONString(encryptedObjStr string) (string, error) {
+	// Unmarshal the encrypted object
+	cryptoObj := cryptoObj{}
+
+	if err := json.Unmarshal([]byte(encryptedObjStr), &cryptoObj); err != nil {
+		return "", fmt.Errorf("failed to unmarshal encrypted data: %v", err)
+	}
+
+	// Decode the base64-encoded encrypted data
+	encryptedData, err := base64.StdEncoding.DecodeString(cryptoObj.EncryptedData)
+	if err != nil {
+		return "", fmt.Errorf("failed to decode base64 data: %v", err)
+	}
+
+	// Decrypt the data
+	decrypted, err := Decrypt(encryptedData)
+	if err != nil {
+		return "", fmt.Errorf("failed to decrypt data: %v", err)
+	}
+
+	return string(decrypted), nil
 }
