+++ server/internal/handlers/job.go
@@ -15,6 +15,7 @@ import (
 	"github.com/beego/beego/v2/core/logs"
 	"github.com/beego/beego/v2/server/web"
 	"github.com/datazip/olake-frontend/server/internal/constants"
+	"github.com/datazip/olake-frontend/server/internal/crypto"
 	"github.com/datazip/olake-frontend/server/internal/database"
 	"github.com/datazip/olake-frontend/server/internal/docker"
 	"github.com/datazip/olake-frontend/server/internal/models"
@@ -527,7 +528,7 @@ func (c *JobHandler) GetTaskLogs() {
 // getOrCreateSource finds or creates a source based on the provided config
 func (c *JobHandler) getOrCreateSource(config models.JobSourceConfig, projectIDStr string) (*models.Source, error) {
 	// Encrypt the source configuration
-	encryptedConfig, err := EncryptJSONString(config.Config)
+	encryptedConfig, err := crypto.EncryptJSONString(config.Config)
 	if err != nil {
 		return nil, fmt.Errorf("failed to encrypt source config: %w", err)
 	}
@@ -581,7 +582,7 @@ func (c *JobHandler) getOrCreateSource(config models.JobSourceConfig, projectIDS
 // getOrCreateDestination finds or creates a destination based on the provided config
 func (c *JobHandler) getOrCreateDestination(config models.JobDestinationConfig, projectIDStr string) (*models.Destination, error) {
 	// Encrypt the destination configuration
-	encryptedConfig, err := EncryptJSONString(config.Config)
+	encryptedConfig, err := crypto.EncryptJSONString(config.Config)
 	if err != nil {
 		return nil, fmt.Errorf("failed to encrypt destination config: %w", err)
 	}
